BlockFix
Fixes the implementation of the block so that upon player death the block is reset. Uses a specialised bool to determine whether the block input has been released or not, this is passed through to the character where the relevant logic is followed-through instead of using a flip-flop or multi-gate node.