Custom Kill Volume and Level Edits
Initial Spawn now coincides with the location of the Gondola.
UE4 Kill Volume has been removed and replaced with my own, so that the player is killed via damage and not just destroyed.
An example checkpoint has been added at the top of the stairs just before the factory location.